fix(opnsense): migrate string-to-int parsing to int64 for 32-bit safety (upstream #81, extended)
parseStringToInt, safeAtoi and parseLineRateBits now return int64 so
>2^31 counters (10G line rates, interface byte counters, unbound
cumulative stats) survive armv7/386 source builds.

+++ internal/collector/unbound_dns.go
@@ -331,7 +331,7 @@ func (c *unboundDNSCollector) Update(client *opnsense.Client, ch chan<- promethe
 	}
 
 	// Queries by protocol
-	protocols := map[string]int{
+	protocols := map[string]int64{
 		"tcp":    data.QueryTCP,
 		"tcpout": data.QueryTCPOut,
 		"udpout": data.QueryUDPOut,
@@ -401,7 +401,7 @@ func (c *unboundDNSCollector) Update(client *opnsense.Client, ch chan<- promethe
 	)
 
 	// Cache counts
-	caches := map[string]int{
+	caches := map[string]int64{
 		"rrset":   data.CacheRrsetCount,
 		"message": data.CacheMessageCount,
 		"infra":   data.CacheInfraCount,
@@ -415,7 +415,7 @@ func (c *unboundDNSCollector) Update(client *opnsense.Client, ch chan<- promethe
 	}
 
 	// Memory bytes
-	memComponents := map[string]int{
+	memComponents := map[string]int64{
 		"rrset_cache":   data.MemCacheRrset,
 		"message_cache": data.MemCacheMessage,
 		"iterator":      data.MemModIterator,
